# src/HouseholdLedger.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HouseholdLedger.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/months/{year}/{month}/budget-plan": {
      "get": {
        "tags": [
          "MonthlyBudgetPlan"
        ],
        "summary": "Gets the saved plan for a calendar month.",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The calendar year.",
            "required": true,
            "schema": {
              "maximum": 9999,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "The calendar month.",
            "required": true,
            "schema": {
              "maximum": 12,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyBudgetPlanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MonthlyBudgetPlan"
        ],
        "summary": "Creates or deliberately revises the plan for a calendar month.",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The calendar year.",
            "required": true,
            "schema": {
              "maximum": 9999,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "The calendar month.",
            "required": true,
            "schema": {
              "maximum": 12,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "description": "The cancellation token.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonthlyBudgetPlanRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MonthlyBudgetPlanRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MonthlyBudgetPlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyBudgetPlanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/months/{year}/{month}/budget-review": {
      "get": {
        "tags": [
          "MonthlyBudgetPlan"
        ],
        "summary": "Reviews a month using its saved intention and actual ledger activity.",
        "parameters": [
          {
            "name": "year",
            "in": "path",
            "description": "The calendar year.",
            "required": true,
            "schema": {
              "maximum": 9999,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "month",
            "in": "path",
            "description": "The calendar month.",
            "required": true,
            "schema": {
              "maximum": 12,
              "minimum": 1,
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The actual-versus-plan review.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthlyBudgetReviewResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MonthlyBudgetPlanRequest": {
        "required": [
          "expectedIncome",
          "intendedSavings",
          "necessities",
          "optional",
          "culture",
          "unexpected"
        ],
        "type": "object",
        "properties": {
          "expectedIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Expected take-home income.",
            "format": "double"
          },
          "intendedSavings": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The amount the user intends to save.",
            "format": "double"
          },
          "necessities": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Necessities spending.",
            "format": "double"
          },
          "optional": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Optional spending.",
            "format": "double"
          },
          "culture": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Culture spending.",
            "format": "double"
          },
          "unexpected": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Unexpected spending.",
            "format": "double"
          }
        },
        "description": "Supplies the user-entered amounts for one calendar month's budget plan."
      },
      "MonthlyBudgetPlanResponse": {
        "required": [
          "year",
          "month",
          "expectedIncome",
          "intendedSavings",
          "necessities",
          "optional",
          "culture",
          "unexpected",
          "lastRevisedAt"
        ],
        "type": "object",
        "properties": {
          "year": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The calendar year.",
            "format": "int32"
          },
          "month": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "The calendar month.",
            "format": "int32"
          },
          "expectedIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Expected take-home income.",
            "format": "double"
          },
          "intendedSavings": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The intended savings amount.",
            "format": "double"
          },
          "necessities": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Necessities spending.",
            "format": "double"
          },
          "optional": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Optional spending.",
            "format": "double"
          },
          "culture": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Culture spending.",
            "format": "double"
          },
          "unexpected": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Planned Unexpected spending.",
            "format": "double"
          },
          "lastRevisedAt": {
            "type": "string",
            "description": "When the plan was last saved.",
            "format": "date-time"
          }
        },
        "description": "Describes a saved monthly plan."
      },
      "MonthlyBudgetReviewResponse": {
        "required": [
          "expectedIncome",
          "intendedSavings",
          "actualIncome",
          "incomeVariance",
          "unallocatedRemainder",
          "lastRevisedAt",
          "classifications"
        ],
        "type": "object",
        "properties": {
          "expectedIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "description": "Expected income, or null when no plan exists.",
            "format": "double"
          },
          "intendedSavings": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "description": "Intended savings, or null when no plan exists.",
            "format": "double"
          },
          "actualIncome": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Income explicitly confirmed as received.",
            "format": "double"
          },
          "incomeVariance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "description": "Actual minus expected income, or null when no plan exists.",
            "format": "double"
          },
          "unallocatedRemainder": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Positive confirmed income remaining after recorded expenses.",
            "format": "double"
          },
          "lastRevisedAt": {
            "type": [
              "null",
              "string"
            ],
            "description": "When the plan was last saved, or null when no plan exists.",
            "format": "date-time"
          },
          "classifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MonthlyClassificationReviewResponse"
            },
            "description": "Planned and actual spending comparisons."
          },
          "cashflowDifference": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Signed confirmed income minus recorded expenses; not a balance or savings claim.",
            "format": "double",
            "default": 0
          }
        },
        "description": "Summarizes saved monthly intentions and authoritative ledger actuals."
      },
      "MonthlyClassificationReviewResponse": {
        "required": [
          "classification",
          "plannedAmount",
          "actualAmount",
          "variance"
        ],
        "type": "object",
        "properties": {
          "classification": {
            "type": "string",
            "description": "The existing Kakeibo-inspired classification name."
          },
          "plannedAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The amount planned, or zero when no plan exists.",
            "format": "double"
          },
          "actualAmount": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "The amount of persisted expenses.",
            "format": "double"
          },
          "variance": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "number",
              "string"
            ],
            "description": "Actual minus planned spending.",
            "format": "double"
          }
        },
        "description": "Compares planned spending to recorded spending for a classification."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  },
...
}